// src/app/generated/steamfitter.api/model/scenarioMembership.ts
/**
 * Steamfitter API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ScenarioMembership { 
    id?: string;
    scenarioId?: string;
    userId?: string | null;
    groupId?: string | null;
    roleId?: string;
}